+++ inc/settings-class.php
@@ -16,11 +16,15 @@ class EJO_Testimonials_Settings
 	//* Plugin setup.
 	public function __construct() 
 	{
+		//* Add Settings Page
+		add_action( 'admin_menu', array( $this, 'add_testimonials_setting_menu' ) );
+
 		//* Register Settings for Settings Page
 		add_action( 'admin_init', array( $this, 'initialize_testimonials_settings' ) );
 
-		//* Add Settings Page
-		add_action( 'admin_menu', array( $this, 'add_testimonials_setting_menu' ) );
+		//* Save settings (before init, because post type registers on init)
+		//* I probably should be using Settings API..
+		add_action( 'init', array( $this, 'save_testimonials_settings' ), 1 );
 
 		//* Add scripts to settings page
 		add_action( 'admin_enqueue_scripts', array( $this, 'add_testimonials_settings_scripts_and_styles' ) ); 
@@ -52,6 +56,22 @@ class EJO_Testimonials_Settings
 		} 
 	}
 
+	//* Save testimonials settings
+	public function save_testimonials_settings()
+	{
+		if (isset($_POST['submit']) && !empty($_POST['ejo-testimonials-settings']) ) {
+
+			//* Escape slug
+			$_POST['ejo-testimonials-settings']['slug'] = sanitize_title( $_POST['ejo-testimonials-settings']['slug'] );
+
+			//* Strip slashes
+			$_POST['ejo-testimonials-settings']['description'] = stripslashes( $_POST['ejo-testimonials-settings']['description'] );
+
+			//* Update settings
+			update_option( "ejo_testimonials_settings", $_POST['ejo-testimonials-settings'] ); 				
+		}
+	}
+
 	//*
 	public function testimonials_settings()
 	{
@@ -60,12 +80,10 @@ class EJO_Testimonials_Settings
 			<h2>Referentie Instellingen</h2>
 
 			<?php 
-			// Save testimonials data
-			if (isset($_POST['submit']) ) {
+			//* Let user know the settings are saved
+			if (isset($_POST['submit']) && !empty($_POST['ejo-testimonials-settings']) ) {
 
-				if (!empty($_POST['ejo-testimonials-settings'])) {
-					update_option( "ejo_testimonials_settings", $_POST['ejo-testimonials-settings'] ); 
-				}
+				flush_rewrite_rules(); //* Flush rewrite rules because archive slug could have changed
 
 				echo "<div class='updated'><p>Testimonial settings updated successfully.</p></div>";
 			}
@@ -78,12 +96,13 @@ class EJO_Testimonials_Settings
 
 				<?php submit_button( 'Wijzigingen opslaan' ); ?>
 				<?php // submit_button( 'Standaard Instellingen', 'secondary', 'reset' ); ?>
-			
+
 			</form>
 
 		</div>
 	<?php
 	}
+
 
     public function show_testimonials_settings() 
     {
@@ -93,46 +112,64 @@ class EJO_Testimonials_Settings
     	//* Linktext
 		// $linktext = (isset($ejo_testimonials_settings['linktext'])) ? $ejo_testimonials_settings['linktext'] : 'Lees Meer';
 
-		//* Archive
-		$archive = (isset($ejo_testimonials_settings['archive'])) ? $ejo_testimonials_settings['archive'] : 'testimonials';
+		//* Archive title
+		$title = (!empty($ejo_testimonials_settings['title'])) ? $ejo_testimonials_settings['title'] : EJO_Testimonials::$post_type_name;
 
+		//* Archive description
+		$description = (!empty($ejo_testimonials_settings['description'])) ? $ejo_testimonials_settings['description'] : '';
+
+		//* Archive slug
+		$slug = (!empty($ejo_testimonials_settings['slug'])) ? $ejo_testimonials_settings['slug'] : EJO_Testimonials::$slug;
+		
     	?>
     	<table class="form-table">
 			<tbody>
 
-				<?php 
-				/*
 				<tr>					
-					<th scope="row" style="width: 140px">
-						<label for="ejo-testimonials-settings-linktext">Linktekst</label>
+					<th scope="row">
+						<label for="ejo-testimonials-settings-title">Title</label>
 					</th>
 					<td>
 						<input
-							id="ejo-testimonials-settings-linktext"
-							value="<?php echo $linktext; ?>"
+							id="ejo-testimonials-settings-title"
+							value="<?php echo $title; ?>"
 							type="text"
-							name="ejo-testimonials-settings[linktext]"
+							name="ejo-testimonials-settings[title]"
 							class="text"
+							style="width"
 						>
-						<span class="description">Tekst op de button wanneer er gelinkt wordt naar een referentie.</span>
+						<p class="description">Wordt getoond op de archiefpagina, breadcrumbs en meta's tenzij anders aangegeven</p>
 					</td>
 				</tr>
-				*/
-				?>
 
 				<tr>					
-					<th scope="row" style="width: 140px">
-						<label for="ejo-testimonials-settings-archive">Archief url</label>
+					<th scope="row">
+						<label for="ejo-testimonials-settings-description">Beschrijving</label>
+					</th>
+					<td>
+						<textarea
+							id="ejo-testimonials-settings-description"
+							name="ejo-testimonials-settings[description]"
+							class="text"
+						><?php echo $description; ?></textarea>
+						<p class="description">De beschrijving kan getoond worden op de archiefpagina (afhankelijk van het thema)</p>
+					</td>
+				</tr>
+
+				<tr>					
+					<th scope="row">
+						<label for="ejo-testimonials-settings-slug">Slug</label>
 					</th>
 					<td>
 						<input
-							id="ejo-testimonials-settings-archive"
-							value="<?php echo $archive; ?>"
+							id="ejo-testimonials-settings-slug"
+							value="<?php echo $slug; ?>"
 							type="text"
-							name="ejo-testimonials-settings[archive]"
+							name="ejo-testimonials-settings[slug]"
 							class="text"
+							style="width"
 						>
-						<span class="description">slug voor archief</span>
+						<p class="description">Bepaalt de url van de archiefpagina</p>
 					</td>
 				</tr>