+++ CHANGELOG.md
@@ -1,2 +1,12 @@
-### 04-07-2015: 0.9.3
-Fallback image if no thumbnail is present
+### Changelog
+
+1.1
+- Added option to change archive title and description
+- Flush rewrite when slug is changed
+- Minor organizational changes
+
+1.0
+- Complete rewrite. Removed most output, let theme handle that
+
+0.9.3
+- Fallback image if no thumbnail is present